third_party/blink/web_tests/editing/pasteboard/do-not-copy-unnecessary-styles-expected.txt


<div>
This test verifies that WebKit does not erroneously clone nodes hierarchy when copying text with background color.
To manually this test, select the text inside the div with the border, cut and paste it back. You should not see an additional red border.
<div style="border: 2px solid red">
<div id="test"><span style="background-color: rgb(187, 187, 187);">Select this text</span><br></div>
</div>
</div>
